Mudei o repositório de IA para a camada correta: iaAnalyzeFeedbackRepository.ts
Ajustei o service para consumir esse repositório:

Import novo em iaAnalyzeService.ts:3
Uso no fluxo em iaAnalyzeService.ts:258
Removi o arquivo antigo que estava na pasta errada:
backends/api-gateway/src/server/express/controllers/services/iaAnalyze/iaAnalyzeFeedbackRepository.ts

Mantive o endpoint fino apontando para handler:
EndpointsIAAnalyze.ts:1

Validação:

Lint do API Gateway: OK
Typecheck do API Gateway: OK
Sem erros nos arquivos alterados
